Moved policy rollouts into separate file
