+++ src/distance/rollout.py
