+++ src/experiments/shaping.py
@@ -4,7 +4,7 @@ import json
 from env import RandomEnv
 from env.reward import random_reward, interpolate
 from distance.canon import canon_and_norm
-from policy import optimize, policy_return, policy_returns
+from distance.rollout import optimize, policy_return, policy_returns
 from env import RandomEnv
 from distance.distance_funcs import epic, epic_torch
 from distance.shaping import shaped_distance, maximal_transition_dist
